+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/sim/loaders/urdf_loader.py
@@ -9,18 +9,14 @@ import json
 import os
 import pathlib
 import random
-from dataclasses import MISSING
 from datetime import datetime
-from typing import Optional
 
 import omni.kit.commands
 from omni.isaac.urdf import _urdf as omni_urdf
 
-from omni.isaac.orbit.utils import configclass
 from omni.isaac.orbit.utils.io import dump_yaml
 
-__all__ = ["UrdfLoaderCfg", "UrdfLoader"]
-
+from .urdf_loader_cfg import UrdfLoaderCfg
 
 _DRIVE_TYPE = {
     "none": 0,
@@ -36,75 +32,6 @@ _NORMALS_DIVISION = {
     "none": 3,
 }
 """Mapping from normals division name to urdf importer normals division number."""
-
-
-@configclass
-class UrdfLoaderCfg:
-    """The configuration class for UrdfLoader."""
-
-    urdf_path: str = MISSING
-    """The path to the urdf file (e.g. path/to/urdf/robot.urdf)."""
-
-    usd_dir: Optional[str] = None
-    """The output directory path to store the generated USD file. Defaults to :obj:`None`.
-
-    If set to :obj:`None`, it is resolved as ``/tmp/Orbit/usd_{date}_{time}_{random}``, where
-    the parameters in braces are runtime generated.
-    """
-
-    usd_file_name: Optional[str] = None
-    """The name of the generated usd file. Defaults to :obj:`None`.
-
-    If set to :obj:`None`, it is resolved from the urdf file name.
-    """
-
-    force_usd_conversion: bool = False
-    """Force the conversion of the urdf file to usd. Defaults to False."""
-
-    link_density = 0.0
-    """Default density used for links. Defaults to 0.
-
-    This setting is only effective if ``"inertial"`` properties are missing in the URDF.
-    """
-
-    import_inertia_tensor: bool = True
-    """Import the inertia tensor from urdf. Defaults to True.
-
-    If the ``"inertial"`` tag is missing, then it is imported as an identity.
-    """
-
-    convex_decompose_mesh = False
-    """Decompose a convex mesh into smaller pieces for a closer fit. Defaults to False."""
-
-    fix_base: bool = MISSING
-    """Create a fix joint to the root/base link. Defaults to True."""
-
-    merge_fixed_joints: bool = False
-    """Consolidate links that are connected by fixed joints. Defaults to False."""
-
-    self_collision: bool = False
-    """Activate self-collisions between links of the articulation. Defaults to False."""
-
-    default_drive_type: str = "none"
-    """The drive type used for joints. Defaults to ``"none"``.
-
-    The drive type dictates the loaded joint PD gains and USD attributes for joint control:
-
-    * ``"none"``: The joint stiffness and damping are set to 0.0.
-    * ``"position"``: The joint stiff and damping are set based on the URDF file or provided configuration.
-    * ``"velocity"``: The joint stiff is set to zero and damping is based on the URDF file or provided configuration.
-    """
-
-    default_drive_stiffness: float = 0.0
-    """The default stiffness of the joint drive. Defaults to 0.0."""
-
-    default_drive_damping: float = 0.0
-    """The default damping of the joint drive. Defaults to 0.0.
-
-    Note:
-        If set to zero, the values parsed from the URDF joint tag ``"<dynamics><damping>"`` are used.
-        Otherwise, it is overridden by the configured value.
-    """
 
 
 class UrdfLoader:
@@ -221,8 +148,13 @@ class UrdfLoader:
 
     @property
     def usd_instanceable_meshes_path(self) -> str:
-        """The path to the USD mesh file."""
-        return os.path.join(self.usd_dir, "Props", "instanceable_meshes.usd")
+        """The path to the USD mesh file.
+
+        This is a relative path with respect to the USD directory. This is to ensure that the mesh references
+        in the generated USD file are resolved relatively, which is important when the USD file is moved to
+        a different location.
+        """
+        return os.path.join(".", "Props", "instanceable_meshes.usd")
 
     """
     Private helpers.
@@ -242,7 +174,7 @@ class UrdfLoader:
 
         # -- instancing settings
         # meshes will be placed in a separate usd file
-        import_config.set_make_instanceable(True)
+        import_config.set_make_instanceable(cfg.make_instanceable)
         import_config.set_instanceable_usd_path(self.usd_instanceable_meshes_path)
 
         # -- asset settings